requirements.txt
fastapi
uvicorn[standard]
httpx
requests
pydantic-settings
openai>=1.0.0
msal                # Microsoft auth library
cryptography        # token-encryption helper
